refactor: PostRepository를 인터페이스로 추출, InMemoryPostRepository 구현체 선언
